requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
discogs-client==2.2.1
Django==2.0.4
django-cleanup==2.1.0
django-sslserver==0.20
djangorestframework==3.8.2
idna==2.6
mysqlclient==1.3.12
oauthlib==2.0.7
Pillow==5.1.0
pycparser==2.18
pyOpenSSL==17.5.0
python-dotenv==0.8.2
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
